==> 3 - Combine predictions.ipynb <==
import pandas as pd

scores = pd.merge(pd.read_csv('lgb_scores.csv'), pd.read_csv('nn_scores.csv'))
scores['score'] = scores.lgb_score + 0.5 * scores.nn_score
scores.head()

sub = scores.groupby('context_id').apply(
    lambda x: x.sort_values('score', ascending=False).reply_id
).reset_index(level=0)

sub.to_csv('combined-final-sub.tsv', index=False, header=False, sep='\t')
